requirements.txt
Flask>=2.3.3
python-dotenv>=1.0.1
google-genai>=0.3.0
PyPDF2>=3.0.1
